Refactor: Move hooks into their own files
- some hooks have been renamed to be more specific. Example:  the
"Chart" hook is now StatsDashboardHook
- There are a few things in app.js that don't see to be used - I've
tagged them with `REVIEW` comments.
