+++ assets/js/hooks/alpine.js
@@ -1,0 +1,12 @@
+export default {
+  mounted() {
+    this.el.dispatchEvent(new CustomEvent('mounted', {
+      bubbles: true
+    }));
+  },
+  updated() {
+    this.el.dispatchEvent(new CustomEvent('updated', {
+      bubbles: true
+    }));
+  },
+}
