+++ assets/js/hooks/index.js
@@ -1,0 +1,31 @@
+import TasksList from "./tasks_list";
+import RidersList from "./riders_list";
+import ConversationList from "./conversation_list";
+import MessageList from "./message_list";
+import RiderSelectionList from "./rider_selection_list";
+import Alpine from "./alpine";
+import Autocomplete from './autocomplete';
+import EmojiButtonHook from './emoji_button';
+import CheckboxAll from './checkbox_all';
+import StatsDashboardChart from './stats_dashboard_chart';
+import FrameHook from './frame';
+import TagsComponentHook from './tags_component';
+import LeafletMap from './leaflet_map';
+
+
+export default {
+  TasksList,
+  RidersList,
+  ConversationList,
+  MessageList,
+  RiderSelectionList,
+  Alpine,
+  Autocomplete,
+  EmojiButtonHook,
+  StatsDashboardChart,
+  CheckboxAll,
+  FrameHook,
+  // REVIEW: this hook is used on TagsComponent module, which isn't used anywhere. Remove it?
+  TagsComponentHook,
+  LeafletMap
+}
