+++ assets/js/hooks/stats_dashboard_chart.js
@@ -1,0 +1,21 @@
+import Chart from 'chart.js/auto';
+
+export default {
+  mounted() {
+    this.chart = new Chart(this.el, {
+      type: 'line',
+      options: {
+        scales: {
+          y: {
+            beginAtZero: true
+          }
+        }
+      }
+    });
+
+    this.handleEvent("update_chart", (data) => {
+      this.chart.data = data;
+      this.chart.update();
+    })
+  }
+}
